refactor(duplex): move pure transforms to coverage-gated stages_duplex_provider.go (#1527)
Splits the pure, side-effect-free transforms (chunkToElement, buildAssistantParts,
takeReasoning) out of the _integration file (which the pre-commit coverage gate
exempts) into stages_duplex_provider.go, where the 80% gate applies. The suffix had
been masking real debt: chunkToElement was 79.5% covered. Added a table-driven test
covering its branches (media defaults, tool calls, interruption, turn_id/transcription)
brings the gated file to ~99%. Session I/O (Process, forward*, send*) stays in the
_integration file. Also de-duplicated the relocated string literals into consts.
